refactor(ruby/x402): drop client, isolate interop fixture under X402::Interop
Per maintainer feedback on #127:
 - "Why do we have a ruby client? We should only support ruby server."
 - "What is the interop code doing here?"

Drop the Ruby x402 client surface entirely. The cross-language harness
exercises Ruby in the server role only; the client side is covered by
the TS/Rust/Go/Python adapters.

Move the remaining x402 interop fixture out of the production-looking
`ruby/lib/x402/` mainline and into `ruby/lib/x402/interop/` so the
fixture-only nature is obvious in the file path:

 - Delete `ruby/lib/x402/client.rb`,
   `ruby/bin/x402-interop-client`,
   `ruby/test/x402_interop_client_test.rb`.
 - Move `ruby/lib/x402/server.rb` -> `ruby/lib/x402/interop/server.rb`.
 - Move `ruby/lib/x402/exact.rb` -> `ruby/lib/x402/interop/exact.rb`.
 - Update `ruby/lib/x402.rb` to require only the interop modules and
   document that the production x402 server surface is out of scope for
   this PR.
 - Update `ruby/bin/x402-interop-server` and
   `ruby/test/x402_interop_server_test.rb` to the new require paths.

Tests after this commit: 186 runs, 680 assertions, 0 failures.
