--- ruby/bin/x402-interop-client
@@ -1,110 +1,0 @@
-#!/usr/bin/env ruby
-# frozen_string_literal: true
-
-require "json"
-require "net/http"
-require "uri"
-
-$LOAD_PATH.unshift(File.expand_path("../lib", __dir__))
-require "x402/client"
-require "x402/exact"
-
-target_url = ENV.fetch("X402_INTEROP_TARGET_URL")
-response = Net::HTTP.get_response(URI(target_url))
-
-headers = {}
-response.each_header do |key, value|
-  headers[key] = value
-end
-selected_requirement, resource = X402::Interop::Client.select_svm_challenge(
-  headers: headers,
-  body: response.body,
-  network: ENV.fetch("X402_INTEROP_NETWORK", "solana:EtWTRABZaYq6iMfeYKouRu166VU2xqa1"),
-  scheme: ENV.fetch("X402_INTEROP_SCHEME", "exact"),
-  preferred_currencies: ENV.fetch("X402_INTEROP_PREFER_CURRENCIES", "")
-    .split(",")
-    .map(&:strip)
-    .reject(&:empty?)
-)
-scheme = ENV.fetch("X402_INTEROP_SCHEME", "exact")
-error_domain = ENV.fetch("X402_INTEROP_INTENT", scheme)
-
-if response.code.to_i == 402 &&
-   !ENV.key?("X402_INTEROP_INTENT") &&
-   scheme == "exact" &&
-   selected_requirement &&
-   ENV["X402_INTEROP_CLIENT_SECRET_KEY"] &&
-   ENV["X402_INTEROP_RPC_URL"]
-  begin
-    payment_signature = X402::Interop::Exact.build_exact_payment_signature_from_rpc(
-      requirement: selected_requirement,
-      client_secret_key: ENV.fetch("X402_INTEROP_CLIENT_SECRET_KEY"),
-      rpc_url: ENV.fetch("X402_INTEROP_RPC_URL"),
-      resource: resource
-    )
-    uri = URI(target_url)
-    paid_request = Net::HTTP::Get.new(uri)
-    paid_request["PAYMENT-SIGNATURE"] = payment_signature
-    paid_response = Net::HTTP.start(uri.hostname, uri.port, use_ssl: uri.scheme == "https") do |http|
-      http.request(paid_request)
-    end
-    paid_headers = {}
-    paid_response.each_header do |key, value|
-      paid_headers[key] = value
-    end
-    paid_body = begin
-      JSON.parse(paid_response.body)
-    rescue JSON::ParserError
-      paid_response.body
-    end
-
-    puts JSON.generate(
-      type: "result",
-      implementation: "ruby",
-      role: "client",
-      ok: paid_response.code.to_i.between?(200, 299),
-      status: paid_response.code.to_i,
-      responseHeaders: paid_headers,
-      responseBody: paid_body,
-      settlement: X402::Interop::Client.header_value(
-        paid_headers,
-        ENV.fetch("X402_INTEROP_SETTLEMENT_HEADER", "x-fixture-settlement")
-      )
-    )
-    exit 0
-  rescue StandardError => e
-    puts JSON.generate(
-      type: "result",
-      implementation: "ruby",
-      role: "client",
-      ok: false,
-      status: response.code.to_i,
-      responseHeaders: headers,
-      responseBody: {
-        error: "ruby_exact_client_payment_failed",
-        message: e.message,
-        challengeStatus: response.code.to_i,
-        challengeBody: response.body,
-        selectedRequirement: selected_requirement
-      },
-      settlement: nil
-    )
-    exit 0
-  end
-end
-
-puts JSON.generate(
-  type: "result",
-  implementation: "ruby",
-  role: "client",
-  ok: false,
-  status: response.code.to_i,
-  responseHeaders: headers,
-  responseBody: {
-    error: "ruby_#{error_domain}_client_not_implemented",
-    challengeStatus: response.code.to_i,
-    challengeBody: response.body,
-    selectedRequirement: selected_requirement
-  },
-  settlement: nil
-)
